move peer fns to coordinator
